src/main/python/requirements.txt
flask
torch
onnxruntime-openvino
numpy
yolov5
Pillow